+++ src/pygeoprocessing/routing/routing.pyx
@@ -2122,7 +2122,8 @@ def flow_accumulation_mfd(
 
 def distance_to_channel_d8(
         flow_dir_d8_raster_path_band, channel_raster_path_band,
-        target_distance_to_channel_raster_path):
+        target_distance_to_channel_raster_path,
+        weight_raster_path_band=None):
     """Calculate distance to channel with D8 flow.
 
     Parameters:
@@ -2141,6 +2142,11 @@ def distance_to_channel_d8(
         target_distance_to_channel_raster_path (string): path to a raster
             created by this call that has per-pixel distances from a given
             pixel to the nearest downhill channel.
+        weight_raster_path_band (tuple): optional path and band number to a
+            raster that will be used as the per-pixel flow distance
+            weight. If `None`, 1 is the default distance between neighboring
+            pixels. This raster must be the same dimensions as
+            `flow_dir_mfd_raster_path_band`.
 
     Returns:
         None.
@@ -2161,9 +2167,21 @@ def distance_to_channel_d8(
     # properties of the parallel rasters
     cdef int raster_x_size, raster_y_size
 
+    # these area used to store custom per-pixel weights and per-pixel values
+    # for distance updates
+    cdef float weight_val, pixel_val
+
     # used for time-delayed logging
     cdef time_t last_log_time
     last_log_time = ctime(NULL)
+
+    for path in (
+            flow_dir_d8_raster_path_band, channel_raster_path_band,
+            weight_raster_path_band):
+        if path is not None and not _is_raster_path_band_formatted(path):
+            raise ValueError(
+                "%s is supposed to be a raster band tuple but it's not." % (
+                    path))
 
     distance_nodata = -1
     pygeoprocessing.new_raster_from_base(
@@ -2174,6 +2192,11 @@ def distance_to_channel_d8(
         gtiff_creation_options=GTIFF_CREATION_OPTIONS)
     distance_to_channel_managed_raster = _ManagedRaster(
         target_distance_to_channel_raster_path, 1, 1)
+
+    cdef _ManagedRaster weight_raster = None
+    if weight_raster_path_band:
+        weight_raster = _ManagedRaster(
+            weight_raster_path_band[0], weight_raster_path_band[1], 0)
 
     channel_managed_raster = _ManagedRaster(
         channel_raster_path_band[0], channel_raster_path_band[1], 0)
@@ -2227,15 +2250,15 @@ def distance_to_channel_d8(
                     continue
 
                 distance_to_channel_stack.push(
-                    PixelType(0.0, xi+xoff-1, yi+yoff-1, 0.0))
+                    PixelType(0.0, xi+xoff-1, yi+yoff-1, 0))
 
                 while not distance_to_channel_stack.empty():
                     xi_q = distance_to_channel_stack.top().xi
                     yi_q = distance_to_channel_stack.top().yi
-                    dist_q = distance_to_channel_stack.top().value
+                    pixel_val = distance_to_channel_stack.top().value
                     distance_to_channel_stack.pop()
 
-                    distance_to_channel_managed_raster.set(xi_q, yi_q, dist_q)
+                    distance_to_channel_managed_raster.set(xi_q, yi_q, pixel_val)
 
                     for i_n in xrange(8):
                         xi_n = xi_q+NEIGHBOR_OFFSET_ARRAY[2*i_n]
@@ -2252,15 +2275,24 @@ def distance_to_channel_d8(
 
                         if (flow_dir_d8_managed_raster.get(xi_n, yi_n) ==
                                 D8_REVERSE_DIRECTION[i_n]):
+                            # if a weight is passed we use it directly and do
+                            # not consider that a diagonal pixel is further
+                            # away than an adjacent one. If no weight is used
+                            # then "distance" is being calculated and we account
+                            # for diagonal distance.
+                            if weight_raster is not None:
+                                weight_val = weight_raster.get(xi_n, yi_n)
+                            else:
+                                weight_val = (SQRT2 if i_n % 2 else 1)
+
                             distance_to_channel_stack.push(
                                 PixelType(
-                                    SQRT2 + dist_q if i_n % 2 else dist_q + 1,
-                                    xi_n, yi_n, 0.0))
+                                    weight_val + pixel_val, xi_n, yi_n, 0))
 
 
 def distance_to_channel_mfd(
         flow_dir_mfd_raster_path_band, channel_raster_path_band,
-        target_distance_to_channel_raster_path):
+        target_distance_to_channel_raster_path, weight_raster_path_band=None):
     """Calculate distance to channel with multiple flow direction.
 
     Parameters:
@@ -2277,6 +2309,11 @@ def distance_to_channel_mfd(
         target_distance_to_channel_raster_path (string): path to a raster
             created by this call that has per-pixel distances from a given
             pixel to the nearest downhill channel.
+        weight_raster_path_band (tuple): optional path and band number to a
+            raster that will be used as the per-pixel flow distance
+            weight. If `None`, 1 is the default distance between neighboring
+            pixels. This raster must be the same dimensions as
+            `flow_dir_mfd_raster_path_band`.
 
     Returns:
         None.
@@ -2302,9 +2339,21 @@ def distance_to_channel_mfd(
     # properties of the parallel rasters
     cdef int raster_x_size, raster_y_size
 
+    # this value is used to store the current weight which might be 1 or
+    # come from a predefined flow accumulation weight raster
+    cdef double weight_val
+
     # used for time-delayed logging
     cdef time_t last_log_time
     last_log_time = ctime(NULL)
+
+    for path in (
+            flow_dir_mfd_raster_path_band, channel_raster_path_band,
+            weight_raster_path_band):
+        if path is not None and not _is_raster_path_band_formatted(path):
+            raise ValueError(
+                "%s is supposed to be a raster band tuple but it's not." % (
+                    path))
 
     distance_nodata = -1
     pygeoprocessing.new_raster_from_base(
@@ -2328,6 +2377,11 @@ def distance_to_channel_mfd(
         flow_dir_mfd_raster_path_band[0], gdal.OF_RASTER)
     flow_dir_mfd_band = flow_dir_mfd_raster.GetRasterBand(
         flow_dir_mfd_raster_path_band[1])
+
+    cdef _ManagedRaster weight_raster = None
+    if weight_raster_path_band:
+        weight_raster = _ManagedRaster(
+            weight_raster_path_band[0], weight_raster_path_band[1], 0)
 
     flow_dir_raster_info = pygeoprocessing.get_raster_info(
         flow_dir_mfd_raster_path_band[0])
@@ -2431,8 +2485,18 @@ def distance_to_channel_mfd(
                                 FlowPixelType(xi_n, yi_n, 0, 0.0))
                             break
 
+                        # if a weight is passed we use it directly and do
+                        # not consider that a diagonal pixel is further
+                        # away than an adjacent one. If no weight is used
+                        # then "distance" is being calculated and we account
+                        # for diagonal distance.
+                        if weight_raster is not None:
+                            weight_val = weight_raster.get(xi_n, yi_n)
+                        else:
+                            weight_val = (SQRT2 if i_n % 2 else 1)
+
                         pixel.value += flow_dir_weight * (
-                            (SQRT2 if i_n % 2 else 1) + n_distance)
+                            weight_val + n_distance)
 
                     if preempted:
                         continue